refactor: YoutubeCategory 엔티티 별도 패키지로 이동
